refactor(checkpoint): Improved checkpointing behaviour for better UX.
Checkpoints and logs are set separately and managed in different folders. We can combine them into a single "runs" directory for one source of truth.

This makes the restore methodology a little easier too.

+++ velora/disco/config/state.py
@@ -480,7 +480,7 @@ class RuleTrainerState:
                 disco_h_size,
                 meta_h_size,
             ),
-            meta_step=0,
+            meta_step=jnp.int32(0),
         )
 
     def update_opt(self, new_opt_state: optax.OptState) -> Self:
